modularize select coordinates
